backend/requirements.txt
langchain 
pymongo 
azure-ai-textanalytics 
youtube-transcript-api 
google-api-core                         
google-api-python-client
langchain_community
pydantic
openai
langchain-openai               
langchain-mongodb
fastapi[all]
uvicorn
html2text
bs4
uvicorn 
python-dotenv
scrapy
python-multipart 